ai_ecommerce_assistant/modules/recommendations.py
"""
Recommendations module for personalized product suggestions.
"""

from typing import Dict, List, Optional
import numpy as np
from sklearn.preprocessing import StandardScaler
from transformers import AutoTokenizer, AutoModel
import torch

class RecommendationEngine:
    """
    Engine for generating personalized product recommendations.
    Uses a combination of collaborative filtering and content-based filtering.
    """

    def __init__(self, api_key: str, config: Dict):
        """
        Initialize the recommendation engine.

        Args:
            api_key: API key for authentication
            config: Configuration dictionary
        """
        self.api_key = api_key
        self.config = config
        
        # Initialize models
        self.tokenizer = AutoTokenizer.from_pretrained("bert-base-uncased")
        self.model = AutoModel.from_pretrained("bert-base-uncased")
        self.scaler = StandardScaler()
        
        # Load pre-trained embeddings if available
        self.product_embeddings = self._load_product_embeddings()
        self.user_embeddings = self._load_user_embeddings()

    def _load_product_embeddings(self) -> Dict[str, np.ndarray]:
        """
        Load pre-computed product embeddings.
        
        Returns:
            Dictionary mapping product IDs to their embeddings
        """
        # TODO: Implement loading from database or cache
        return {}

    def _load_user_embeddings(self) -> Dict[str, np.ndarray]:
        """
        Load pre-computed user embeddings.
        
        Returns:
            Dictionary mapping user IDs to their embeddings
        """
        # TODO: Implement loading from database or cache
        return {}

    def _get_product_embedding(self, product_data: Dict) -> np.ndarray:
        """
        Generate embedding for a product using its metadata.

        Args:
            product_data: Dictionary containing product information

        Returns:
            Product embedding as numpy array
        """
        # Combine product features
        text = f"{product_data['name']} {product_data.get('description', '')}"
        
        # Tokenize and get BERT embeddings
        inputs = self.tokenizer(text, return_tensors="pt", padding=True, truncation=True)
        with torch.no_grad():
            outputs = self.model(**inputs)
        
        # Use mean pooling of last hidden states
        embeddings = outputs.last_hidden_state.mean(dim=1).numpy()
        return embeddings[0]

    def _get_user_embedding(
        self,
        user_id: str,
        purchase_history: List[Dict]
    ) -> np.ndarray:
        """
        Generate embedding for a user based on their purchase history.

        Args:
            user_id: User ID
            purchase_history: List of purchase records

        Returns:
            User embedding as numpy array
        """
        if user_id in self.user_embeddings:
            return self.user_embeddings[user_id]

        # Combine product embeddings from purchase history
        product_embeddings = []
        for purchase in purchase_history:
            if purchase['product_id'] in self.product_embeddings:
                product_embeddings.append(self.product_embeddings[purchase['product_id']])

        if not product_embeddings:
            return np.zeros(768)  # Default embedding size for BERT

        # Average the product embeddings
        user_embedding = np.mean(product_embeddings, axis=0)
        return user_embedding

    def get_recommendations(
        self,
        user_id: str,
        limit: int = 10,
        context: Optional[Dict] = None
    ) -> List[Dict]:
        """
        Get personalized product recommendations for a user.

        Args:
            user_id: Unique identifier for the user
            limit: Maximum number of recommendations to return
            context: Optional context information (e.g., current page, search query)

        Returns:
            List of recommended products with scores
        """
        # Get user's purchase history
        purchase_history = self._get_purchase_history(user_id)
        
        # Generate user embedding
        user_embedding = self._get_user_embedding(user_id, purchase_history)
        
        # Get candidate products
        candidate_products = self._get_candidate_products(context)
        
        # Calculate similarity scores
        recommendations = []
        for product in candidate_products:
            if product['id'] not in [p['product_id'] for p in purchase_history]:
                product_embedding = self._get_product_embedding(product)
                similarity = np.dot(user_embedding, product_embedding) / (
                    np.linalg.norm(user_embedding) * np.linalg.norm(product_embedding)
                )
                
                recommendations.append({
                    'product_id': product['id'],
                    'name': product['name'],
                    'score': float(similarity),
                    'price': product['price'],
                    'image_url': product.get('image_url'),
                    'category': product.get('category')
                })

        # Sort by similarity score and return top recommendations
        recommendations.sort(key=lambda x: x['score'], reverse=True)
        return recommendations[:limit]

    def _get_purchase_history(self, user_id: str) -> List[Dict]:
        """
        Retrieve user's purchase history.

        Args:
            user_id: User ID

        Returns:
            List of purchase records
        """
        # TODO: Implement actual database query
        return []

    def _get_candidate_products(self, context: Optional[Dict] = None) -> List[Dict]:
        """
        Get candidate products for recommendations.

        Args:
            context: Optional context information

        Returns:
            List of candidate products
        """
        # TODO: Implement actual database query with context filtering
        return []

    def update_product_embeddings(self, product_data: Dict) -> None:
        """
        Update embeddings for a product.

        Args:
            product_data: Product information
        """
        product_id = product_data['id']
        self.product_embeddings[product_id] = self._get_product_embedding(product_data)
        # TODO: Implement persistence to database or cache

    def update_user_embeddings(self, user_id: str, purchase_history: List[Dict]) -> None:
        """
        Update embeddings for a user.

        Args:
            user_id: User ID
            purchase_history: List of purchase records
        """
        self.user_embeddings[user_id] = self._get_user_embedding(user_id, purchase_history)
        # TODO: Implement persistence to database or cache 