pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-ecommerce-assistant"
version = "0.1.0"
description = "A powerful AI-integrated library for e-commerce websites"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "transformers>=4.15.0",
    "torch>=1.10.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "requests>=2.26.0",
    "pydantic>=1.8.0",
    "pillow>=8.3.0",
    "tensorflow>=2.7.0",
    "nltk>=3.6.0",
    "spacy>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.9b0",
    "isort>=5.9.3",
    "mypy>=0.910",
    "flake8>=3.9.2",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=ai_ecommerce_assistant --cov-report=term-missing" 